Issue-127: RestlasticSearchClientTest created
