--- elasticsearch-core/src/test/scala/com/sumologic/elasticsearch/restlastic/RestlasticClientIntegrationTest.scala
@@ -1,5 +1,0 @@
-package com.sumologic.elasticsearch.restlastic
-
-trait RestlasticClientIntegrationTest {
-
-}
